refactor : 최근검색어 subnavbar 에 가려지지 않도록 수정 - 데스크탑용 검색 컴포넌트 popover 적용 - 검색컴포넌트 명칭 수정
